Split bookmark weaver orchestration modules
