+++ apps/extension/src/hooks/useBookmarkWeaver.ts
@@ -1,18 +1,36 @@
-import { useState, useEffect, useCallback, useRef } from 'react';
+import { useState, useEffect, useCallback, useMemo, useRef } from 'react';
 import { BookmarkNode } from '../components/BookmarkTree';
 import { ClusteringSettings, normalizeClusteringSettings } from '../lib/clusteringSettings';
-import { BookmarkRootTitle, ROOT_IDS, ROOT_TITLES } from '../lib/bookmarkImport';
+import { BookmarkRootTitle } from '../lib/bookmarkImport';
 import {
     BookmarkStats,
     StructureAssignment,
     collectDuplicateChromeIds,
-    countBookmarksInTree,
     countDuplicateAssignments,
-    getBookmarkChromeId,
     normalizeBookmarkUrl,
     pruneBookmarksFromTree,
-    summarizeStructure,
 } from '../lib/bookmarkStructure';
+import {
+    BackupClient,
+    deleteStructureVersion as deleteStoredStructureVersion,
+    loadStructureVersions as loadStoredStructureVersions,
+    saveStructureVersion as saveStoredStructureVersion,
+} from '../lib/backupClient';
+import { StructureClient, WeavingProgress } from '../lib/structureClient';
+import { buildStructurePreview } from '../lib/structurePreviewBuilder';
+import { applyChromeBookmarkPlan } from '../lib/chromeApplyPlan';
+import {
+    buildBookmarkRootSnapshot,
+    clearPersistedOverflowBookmarks,
+    collectScannedBookmarks,
+    createEmptyProgress,
+    loadPersistedOverflowBookmarks,
+    persistOverflowBookmarks,
+    savePreOrganizeBackup,
+    ScannedBookmark,
+} from '../lib/processingSession';
+
+export type { BookmarkBackupSnapshot, BookmarkStructureVersion } from '../lib/backupClient';
 
 const BACKEND_URL = (import.meta.env.VITE_BACKEND_URL as string | undefined)?.replace(/\/$/, '') ?? '';
 const BACKEND_UNAVAILABLE_MESSAGE = BACKEND_URL
@@ -22,10 +40,6 @@ const DEAD_LINK_SCAN_REQUEST_TIMEOUT_MS = 45000;
 const AUTO_RENAME_REQUEST_TIMEOUT_MS = 120000;
 const STRUCTURE_REQUEST_TIMEOUT_MS = 45000;
 const DEFAULT_FREE_TIER_LIMIT = 500;
-const STRUCTURE_VERSIONS_STORAGE_KEY = 'bookmarkStructureVersions';
-const PRE_ORGANIZE_BACKUP_KEY = 'preOrganizeBackup';
-const OVERFLOW_BOOKMARKS_STORAGE_KEY = 'bookmarkWeaverOverflowBookmarks';
-const MAX_STRUCTURE_VERSIONS = 20;
 const DEFAULT_ROOT_TITLE: BookmarkRootTitle = 'Other Bookmarks';
 export type AppStatus = 'idle' | 'weaving' | 'ready' | 'done' | 'error' | 'limit_exceeded';
 export type WeavingPhase = 'backup' | 'ingest' | null;
@@ -35,181 +49,11 @@ export type LimitExceededInfo = {
     limit: number;
 };
 
-export type BookmarkStructureVersion = {
-    id: string;
-    createdAt: string;
-    clusters: BookmarkNode[];
-    stats: BookmarkStats;
-    summary: {
-        folders: number;
-        bookmarks: number;
-    };
-};
-
-export type BookmarkBackupSnapshot = {
-    id: string;
-    name: string;
-    createdAt: string;
-    summary: {
-        folders: number;
-        bookmarks: number;
-    };
-};
-
-type ScannedBookmark = {
-    id: string;
-    url: string;
-    title: string;
-};
-
-type BookmarkRootSnapshot = {
-    bookmarkRoots: Record<string, BookmarkRootTitle>;
-    preferredRoots: Record<string, BookmarkRootTitle>;
-    availableRoots: BookmarkRootTitle[];
-};
-
 const isFailedFetchError = (error: unknown) =>
     error instanceof TypeError && error.message.toLowerCase().includes('failed to fetch');
 
 const isAbortError = (error: unknown) =>
     error instanceof DOMException && error.name === 'AbortError';
-
-type WeavingProgress = {
-    pending: number;
-    pendingRaw: number;
-    enriched: number;
-    embedded: number;
-    errored: number;
-    processing: number;
-    remainingToAssign: number;
-    clusters: number;
-    assigned: number;
-    total: number;
-    isIngesting: boolean;
-    ingestProcessed: number;
-    ingestTotal: number;
-    isClusteringActive: boolean;
-};
-
-const createEmptyProgress = (): WeavingProgress => ({
-    pending: 0,
-    pendingRaw: 0,
-    enriched: 0,
-    embedded: 0,
-    errored: 0,
-    processing: 0,
-    remainingToAssign: 0,
-    clusters: 0,
-    assigned: 0,
-    total: 0,
-    isIngesting: false,
-    ingestProcessed: 0,
-    ingestTotal: 0,
-    isClusteringActive: false
-});
-
-const isBookmarkRootTitle = (value: string): value is BookmarkRootTitle =>
-    ROOT_TITLES.includes(value as BookmarkRootTitle);
-
-const IMPORTED_FOLDER_PATTERN = /^Imported(?: \(\d+\))?$/;
-
-const inferPreferredRootFromAncestors = (
-    ancestorTitles: string[],
-    actualRoot: BookmarkRootTitle
-): BookmarkRootTitle => {
-    const inferredRoot = ancestorTitles.find(isBookmarkRootTitle);
-    if (!inferredRoot || inferredRoot === actualRoot) {
-        return actualRoot;
-    }
-
-    const hasImportedAncestor = ancestorTitles.some((title) => IMPORTED_FOLDER_PATTERN.test(title));
-    const isDirectImportedRoot = ancestorTitles[0] === inferredRoot;
-
-    if (hasImportedAncestor || isDirectImportedRoot) {
-        return inferredRoot;
-    }
-
-    return actualRoot;
-};
-
-const buildBookmarkRootSnapshot = (tree: any[]): BookmarkRootSnapshot => {
-    const bookmarkRoots: Record<string, BookmarkRootTitle> = {};
-    const preferredRoots: Record<string, BookmarkRootTitle> = {};
-    const availableRoots: BookmarkRootTitle[] = [];
-    const topLevelNodes = Array.isArray(tree?.[0]?.children) ? tree[0].children : [];
-
-    topLevelNodes.forEach((node: any) => {
-        const rootTitle =
-            ROOT_TITLES.find((candidate) => node.id === ROOT_IDS[candidate] || node.title === candidate) ??
-            (typeof node.title === 'string' && isBookmarkRootTitle(node.title) ? node.title : null);
-
-        if (!rootTitle) {
-            return;
-        }
-
-        availableRoots.push(rootTitle);
-
-        const visit = (entry: any, ancestorTitles: string[] = []) => {
-            if (entry.url) {
-                bookmarkRoots[entry.id] = rootTitle;
-                preferredRoots[entry.id] = inferPreferredRootFromAncestors(ancestorTitles, rootTitle);
-            }
-            if (Array.isArray(entry.children)) {
-                const nextAncestorTitles = entry?.title ? [...ancestorTitles, String(entry.title)] : ancestorTitles;
-                entry.children.forEach((child: any) =>
-                    visit(child, nextAncestorTitles)
-                );
-            }
-        };
-
-        node.children?.forEach((child: any) => visit(child, []));
-    });
-
-    return {
-        bookmarkRoots,
-        preferredRoots,
-        availableRoots,
-    };
-};
-
-const getOverflowStorageKey = (userId: string) => `${OVERFLOW_BOOKMARKS_STORAGE_KEY}:${userId}`;
-
-const persistOverflowBookmarks = async (userId: string, overflowBookmarks: ScannedBookmark[]) => {
-    if (typeof chrome === 'undefined' || !chrome.storage?.local || !userId) return;
-    await chrome.storage.local.set({ [getOverflowStorageKey(userId)]: overflowBookmarks });
-};
-
-const loadPersistedOverflowBookmarks = async (userId: string) => {
-    if (typeof chrome === 'undefined' || !chrome.storage?.local || !userId) {
-        return [] as ScannedBookmark[];
-    }
-
-    const storageResult = await chrome.storage.local.get([getOverflowStorageKey(userId)]);
-    const stored = storageResult[getOverflowStorageKey(userId)];
-    return Array.isArray(stored) ? (stored as ScannedBookmark[]) : [];
-};
-
-const clearPersistedOverflowBookmarks = async (userId: string) => {
-    if (typeof chrome === 'undefined' || !chrome.storage?.local || !userId) return;
-    await chrome.storage.local.remove(getOverflowStorageKey(userId));
-};
-
-const resolvePreviewRoot = (
-    chromeId: string,
-    availableRoots: BookmarkRootTitle[],
-    actualRootMap: Record<string, BookmarkRootTitle>,
-    preferredRootMap: Record<string, BookmarkRootTitle>
-): BookmarkRootTitle => {
-    const actualRoot = actualRootMap[chromeId] ?? DEFAULT_ROOT_TITLE;
-    const preferredRoot = preferredRootMap[chromeId] ?? actualRoot;
-
-    if (preferredRoot === 'Mobile Bookmarks' && !availableRoots.includes('Mobile Bookmarks')) {
-        return 'Other Bookmarks';
-    }
-
-    return preferredRoot;
-};
-
 
 export const useBookmarkWeaver = (
     accountUserId?: string | null,
@@ -282,6 +126,20 @@ export const useBookmarkWeaver = (
         };
     }, [authAccessToken]);
 
+    const structureClient = useMemo(() => new StructureClient({
+        backendUrl: BACKEND_URL,
+        buildAuthHeaders,
+        getAuthHeaders,
+    }), [buildAuthHeaders, getAuthHeaders]);
+
+    const backupClient = useMemo(() => new BackupClient({
+        backendUrl: BACKEND_URL,
+        accountUserId,
+        canSaveAccountBackups,
+        buildAuthHeaders,
+        getAuthHeaders,
+    }), [accountUserId, buildAuthHeaders, canSaveAccountBackups, getAuthHeaders]);
+
     const loadCurrentBookmarkTreeSnapshot = useCallback(async () => {
         if (typeof chrome === 'undefined' || !chrome.bookmarks) {
             return [] as any[];
@@ -320,10 +178,7 @@ export const useBookmarkWeaver = (
             setUserId(resolvedUserId);
 
             try {
-                const res = await fetch(`${BACKEND_URL}/status/${resolvedUserId}`, {
-                    headers: getAuthHeaders()
-                });
-                const data = await res.json();
+                const data = await structureClient.getStatus(resolvedUserId);
                 if (cancelled) return;
 
                 if (data.isPremium) setIsPremium(true);
@@ -364,7 +219,7 @@ export const useBookmarkWeaver = (
         return () => {
             cancelled = true;
         };
-    }, [accountUserId, getAuthHeaders]);
+    }, [accountUserId, structureClient]);
 
     // Polling Effect
     useEffect(() => {
@@ -378,10 +233,7 @@ export const useBookmarkWeaver = (
 
         const interval = setInterval(async () => {
             try {
-                const res = await fetch(`${BACKEND_URL}/status/${userId}`, {
-                    headers: getAuthHeaders()
-                });
-                const data = await res.json();
+                const data = await structureClient.getStatus(userId);
                 
                 if (data.isPremium) setIsPremium(true);
 
@@ -415,11 +267,7 @@ export const useBookmarkWeaver = (
                     (data.clusters === 0 || (data.remainingToAssign ?? 0) > 0)
                 ) {
                     clusterRecoveryTriggered.current = true;
-                    fetch(`${BACKEND_URL}/trigger-clustering/${userId}`, {
-                        method: 'POST',
-                        headers: buildAuthHeaders(),
-                        body: JSON.stringify({ clusteringSettings: effectiveClusteringSettings })
-                    })
+                    structureClient.triggerClustering(userId, effectiveClusteringSettings)
                         .catch((err) => console.error('[WEAVING] Failed to trigger recovery clustering', err));
                 }
 
@@ -437,7 +285,7 @@ export const useBookmarkWeaver = (
         }, 2000);
 
         return () => clearInterval(interval);
-    }, [buildAuthHeaders, effectiveClusteringSettings, getAuthHeaders, status, userId]);
+    }, [effectiveClusteringSettings, status, structureClient, userId]);
     const startWeaving = useCallback(async () => {
         let processingIdentity: { userId: string; accessToken: string };
         try {
@@ -523,16 +371,7 @@ export const useBookmarkWeaver = (
             // 1. Get Bookmarks
             const tree = await loadCurrentBookmarkTreeSnapshot();
 
-            const bookmarks: ScannedBookmark[] = [];
-            const traverse = (node: any) => {
-                if (node.url) {
-                    bookmarks.push({ id: node.id, url: node.url, title: node.title });
-                }
-                if (node.children) {
-                    node.children.forEach(traverse);
-                }
-            };
-            traverse(tree[0]);
+            const bookmarks = collectScannedBookmarks(tree);
             const totalBookmarks = bookmarks.length;
 
             if (!isPremium && totalBookmarks > DEFAULT_FREE_TIER_LIMIT) {
@@ -551,15 +390,8 @@ export const useBookmarkWeaver = (
             setWeavingPhase('backup');
 
             // 1a. Save a local backup before doing anything
-            const backup = {
-                id: crypto.randomUUID(),
-                createdAt: new Date().toISOString(),
-                tree: tree,
-            };
-            if (typeof chrome !== 'undefined' && chrome.storage?.local) {
-                await chrome.storage.local.set({ [PRE_ORGANIZE_BACKUP_KEY]: backup });
-                console.log('[WEAVING] Pre-organize backup saved to chrome.storage.local');
-            }
+            await savePreOrganizeBackup(tree);
+            console.log('[WEAVING] Pre-organize backup saved to chrome.storage.local');
 
             // Move to ingest phase now that backup is complete
             setWeavingPhase('ingest');
@@ -590,10 +422,10 @@ export const useBookmarkWeaver = (
             setStats({ duplicates: duplicateCount, deadLinks: 0 });
 
             // 2. Send to Backend
-            const response = await fetch(`${BACKEND_URL}/ingest`, {
-                method: 'POST',
-                headers: buildAuthHeaders(processingIdentity.accessToken),
-                body: JSON.stringify({ bookmarks, clusteringSettings: effectiveClusteringSettings }),
+            const response = await structureClient.ingest({
+                bookmarks,
+                clusteringSettings: effectiveClusteringSettings,
+                accessToken: processingIdentity.accessToken,
             });
 
             // Handle 402 Payment Required (limit exceeded)
@@ -631,7 +463,7 @@ export const useBookmarkWeaver = (
             setErrorMessage(message);
             setStatus('error');
         }
-    }, [buildAuthHeaders, effectiveClusteringSettings, ensureProcessingIdentity, isPremium]);
+    }, [effectiveClusteringSettings, ensureProcessingIdentity, isPremium, structureClient]);
 
     const continueWithLimitedBookmarks = useCallback(async () => {
         const limit = limitExceededInfo?.limit ?? 500;
@@ -660,10 +492,9 @@ export const useBookmarkWeaver = (
         }));
 
         try {
-            const response = await fetch(`${BACKEND_URL}/ingest`, {
-                method: 'POST',
-                headers: buildAuthHeaders(),
-                body: JSON.stringify({ bookmarks: slicedBookmarks, clusteringSettings: effectiveClusteringSettings }),
+            const response = await structureClient.ingest({
+                bookmarks: slicedBookmarks,
+                clusteringSettings: effectiveClusteringSettings,
             });
 
             if (response.status === 402) {
@@ -693,7 +524,7 @@ export const useBookmarkWeaver = (
             setErrorMessage(message);
             setStatus('error');
         }
-    }, [buildAuthHeaders, effectiveClusteringSettings, limitExceededInfo, userId]);
+    }, [effectiveClusteringSettings, limitExceededInfo, structureClient, userId]);
 
 
 
@@ -736,17 +567,7 @@ export const useBookmarkWeaver = (
         const controller = new AbortController();
         const timeoutId = setTimeout(() => controller.abort(), DEAD_LINK_SCAN_REQUEST_TIMEOUT_MS);
         try {
-            const response = await fetch(`${BACKEND_URL}/dead-links/check`, {
-                method: 'POST',
-                headers: buildAuthHeaders(),
-                signal: controller.signal,
-                body: JSON.stringify({
-                    bookmarks: assignmentsToScan.map((assignment) => ({
-                        chromeId: assignment.chromeId,
-                        url: assignment.url
-                    }))
-                })
-            });
+            const response = await structureClient.scanDeadLinks(assignmentsToScan, controller.signal);
 
             if (!response.ok) {
                 throw new Error(`Dead-link scan failed: ${response.status}`);
@@ -791,7 +612,7 @@ export const useBookmarkWeaver = (
                 setIsScanningDeadLinks(false);
             }
         }
-    }, [buildAuthHeaders, isPremium, structureAssignments]);
+    }, [isPremium, structureAssignments, structureClient]);
 
     const fetchResults = async (idOverride?: string, silent = false) => {
         const targetId = idOverride || userId;
@@ -803,169 +624,20 @@ export const useBookmarkWeaver = (
                 overflowBookmarksRef.current = await loadPersistedOverflowBookmarks(targetId);
             }
 
-            const res = await fetch(`${BACKEND_URL}/structure/${targetId}`, {
-                signal: controller.signal,
-                headers: getAuthHeaders()
-            });
+            const res = await structureClient.fetchStructure(targetId, controller.signal);
             if (!res.ok) {
                 throw new Error(`Structure fetch failed: ${res.status}`);
             }
             const data = await res.json();
 
-            const clusterDefinitions = new Map<string, { id: string; name: string; parentId: string | null }>();
-            const childClusterIds = new Map<string | null, string[]>();
-            const assignmentSummaries: StructureAssignment[] = [];
-            const bookmarksByRootAndCluster = new Map<BookmarkRootTitle, Map<string, BookmarkNode[]>>();
-            const availableRoots = availableRootsRef.current;
-
-            data.clusters.forEach((cluster: any) => {
-                clusterDefinitions.set(cluster.id, {
-                    id: cluster.id,
-                    name: cluster.name,
-                    parentId: cluster.parent_id ?? null,
-                });
-                const parentKey = cluster.parent_id ?? null;
-                const siblings = childClusterIds.get(parentKey);
-                if (siblings) {
-                    siblings.push(cluster.id);
-                } else {
-                    childClusterIds.set(parentKey, [cluster.id]);
-                }
-            });
-
-            data.assignments.forEach((a: any) => {
-                const rawUrl = a.bookmarks?.url;
-                const chromeId = a.bookmarks?.chrome_id;
-                if (typeof rawUrl === 'string' && typeof chromeId === 'string' && rawUrl && chromeId) {
-                    const rootTitle = resolvePreviewRoot(
-                        chromeId,
-                        availableRoots,
-                        bookmarkRootMapRef.current,
-                        bookmarkPreferredRootMapRef.current
-                    );
-                    assignmentSummaries.push({
-                        bookmarkId: a.bookmark_id,
-                        chromeId,
-                        url: rawUrl,
-                        rootTitle,
-                    });
-
-                    const rootAssignments = bookmarksByRootAndCluster.get(rootTitle) ?? new Map<string, BookmarkNode[]>();
-                    const clusterBookmarks = rootAssignments.get(a.cluster_id) ?? [];
-                    clusterBookmarks.push({
-                        id: `bookmark-${a.bookmark_id}`,
-                        title: a.bookmarks.ai_title || a.bookmarks.title,
-                        originalTitle: a.bookmarks.title,
-                        url: rawUrl,
-                        chromeId,
-                        nodeType: 'bookmark',
-                        rootTitle,
-                    });
-                    rootAssignments.set(a.cluster_id, clusterBookmarks);
-                    bookmarksByRootAndCluster.set(rootTitle, rootAssignments);
-                }
-            });
-            const duplicateCount = countDuplicateAssignments(assignmentSummaries);
-
-            const buildClusterNodeForRoot = (clusterId: string, rootTitle: BookmarkRootTitle): BookmarkNode | null => {
-                const cluster = clusterDefinitions.get(clusterId);
-                if (!cluster) return null;
-
-                const childFolders = (childClusterIds.get(clusterId) || [])
-                    .map((childId) => buildClusterNodeForRoot(childId, rootTitle))
-                    .filter((node): node is BookmarkNode => Boolean(node));
-                const directBookmarks = bookmarksByRootAndCluster.get(rootTitle)?.get(clusterId) || [];
-
-                if (childFolders.length === 0 && directBookmarks.length === 0) {
-                    return null;
-                }
-
-                return {
-                    id: `cluster-${rootTitle}-${clusterId}`,
-                    title: cluster.name,
-                    children: [...childFolders, ...directBookmarks],
-                    parentId: cluster.parentId,
-                    nodeType: 'folder',
-                    rootTitle,
-                };
-            };
-
-            const overflowIds = new Set(overflowBookmarksRef.current.map((bookmark) => bookmark.id));
-            const buildOverflowTree = (nodes: any[]): BookmarkNode[] => {
-                const result: BookmarkNode[] = [];
-
-                for (const node of nodes) {
-                    if (node.url) {
-                        if (!overflowIds.has(node.id)) continue;
-                        const rootTitle = bookmarkRootMapRef.current[node.id] ?? DEFAULT_ROOT_TITLE;
-                        result.push({
-                            id: `overflow-bookmark-${node.id}`,
-                            title: node.title,
-                            originalTitle: node.title,
-                            url: node.url,
-                            chromeId: node.id,
-                            nodeType: 'bookmark',
-                            rootTitle,
-                            isOverflow: true,
-                        });
-                        continue;
-                    }
-
-                    if (!Array.isArray(node.children)) continue;
-
-                    const filteredChildren = buildOverflowTree(node.children);
-                    if (filteredChildren.length === 0) continue;
-
-                    result.push({
-                        id: `overflow-folder-${node.id}`,
-                        title: node.title || 'Untitled Folder',
-                        children: filteredChildren,
-                        nodeType: 'folder',
-                        rootTitle: typeof node.title === 'string' && isBookmarkRootTitle(node.title) ? node.title : undefined,
-                        isOverflow: true,
-                    });
-                }
-
-                return result;
-            };
-
-            const overflowNodes = buildOverflowTree(originalTreeRef.current?.[0]?.children || []);
-            const rootNodes: BookmarkNode[] = [];
-            const overflowCount = overflowBookmarksRef.current.length;
-
-            ROOT_TITLES.forEach((rootTitle) => {
-                const clusterChildren = (childClusterIds.get(null) || [])
-                    .map((clusterId) => buildClusterNodeForRoot(clusterId, rootTitle))
-                    .filter((node): node is BookmarkNode => Boolean(node));
-                const rootChildren = [...clusterChildren];
-
-                if (rootTitle === 'Other Bookmarks' && overflowNodes.length > 0) {
-                    rootChildren.push({
-                        id: 'overflow-unorganized-folder',
-                        title: 'Unorganized Bookmarks',
-                        children: overflowNodes,
-                        nodeType: 'folder',
-                        rootTitle,
-                        isOverflow: true,
-                        badgeLabel: 'Unorganized',
-                    });
-                }
-
-                if (rootChildren.length === 0 && !availableRoots.includes(rootTitle)) {
-                    return;
-                }
-
-                rootNodes.push({
-                    id: `root-${rootTitle}`,
-                    title: rootTitle,
-                    children: rootChildren,
-                    nodeType: 'root',
-                    rootTitle,
-                    badgeLabel:
-                        rootTitle === 'Other Bookmarks' && overflowCount > 0
-                            ? `${overflowCount} extra`
-                            : undefined,
-                });
+            const { rootNodes, assignmentSummaries, duplicateCount } = buildStructurePreview({
+                data,
+                availableRoots: availableRootsRef.current,
+                bookmarkRootMap: bookmarkRootMapRef.current,
+                bookmarkPreferredRootMap: bookmarkPreferredRootMapRef.current,
+                overflowBookmarks: overflowBookmarksRef.current,
+                originalTree: originalTreeRef.current,
+                defaultRootTitle: DEFAULT_ROOT_TITLE,
             });
 
             setClusters(rootNodes);
@@ -1010,12 +682,7 @@ export const useBookmarkWeaver = (
             setIsAutoRenaming(true);
             setErrorMessage(null);
 
-            const response = await fetch(`${BACKEND_URL}/auto-rename/${userId}`, {
-                method: 'POST',
-                headers: buildAuthHeaders(),
-                signal: controller.signal,
-                body: JSON.stringify({ clusteringSettings: effectiveClusteringSettings }),
-            });
+            const response = await structureClient.autoRename(userId, effectiveClusteringSettings, controller.signal);
 
             if (!response.ok) {
                 throw new Error(`Auto rename failed: ${response.status}`);
@@ -1037,7 +704,7 @@ export const useBookmarkWeaver = (
             clearTimeout(timeoutId);
             setIsAutoRenaming(false);
         }
-    }, [buildAuthHeaders, effectiveClusteringSettings, isPremium, userId]);
+    }, [effectiveClusteringSettings, isPremium, structureClient, userId]);
 
     const deleteAllDuplicates = useCallback(async () => {
         if (typeof chrome === 'undefined' || !chrome.bookmarks) return;
@@ -1105,40 +772,11 @@ export const useBookmarkWeaver = (
     }, [isDeletingDeadLinks, updateStateAfterBookmarkRemoval]);
 
     const saveStructureVersion = useCallback(async () => {
-        if (!clusters.length) {
-            throw new Error('No bookmark structure is available to save yet.');
-        }
-
-        const snapshot: BookmarkStructureVersion = {
-            id: crypto.randomUUID(),
-            createdAt: new Date().toISOString(),
-            clusters: JSON.parse(JSON.stringify(clusters)) as BookmarkNode[],
-            stats: { ...stats },
-            summary: summarizeStructure(clusters)
-        };
-
-        if (typeof chrome === 'undefined' || !chrome.storage?.local) {
-            return snapshot;
-        }
-
-        const storageResult = await chrome.storage.local.get([STRUCTURE_VERSIONS_STORAGE_KEY]);
-        const existingVersions = Array.isArray(storageResult[STRUCTURE_VERSIONS_STORAGE_KEY])
-            ? (storageResult[STRUCTURE_VERSIONS_STORAGE_KEY] as BookmarkStructureVersion[])
-            : [];
-        const nextVersions = [snapshot, ...existingVersions].slice(0, MAX_STRUCTURE_VERSIONS);
-        await chrome.storage.local.set({ [STRUCTURE_VERSIONS_STORAGE_KEY]: nextVersions });
-        return snapshot;
+        return saveStoredStructureVersion(clusters, stats);
     }, [clusters, stats]);
 
     const loadStructureVersions = useCallback(async () => {
-        if (typeof chrome === 'undefined' || !chrome.storage?.local) {
-            return [] as BookmarkStructureVersion[];
-        }
-
-        const storageResult = await chrome.storage.local.get([STRUCTURE_VERSIONS_STORAGE_KEY]);
-        return Array.isArray(storageResult[STRUCTURE_VERSIONS_STORAGE_KEY])
-            ? (storageResult[STRUCTURE_VERSIONS_STORAGE_KEY] as BookmarkStructureVersion[])
-            : [];
+        return loadStoredStructureVersions();
     }, []);
 
     const restoreStructureVersion = useCallback(async (versionId: string) => {
@@ -1155,78 +793,27 @@ export const useBookmarkWeaver = (
     }, [loadStructureVersions]);
 
     const deleteStructureVersion = useCallback(async (versionId: string) => {
-        if (typeof chrome === 'undefined' || !chrome.storage?.local) {
-            return;
-        }
-
-        const versions = await loadStructureVersions();
-        const remaining = versions.filter((item) => item.id !== versionId);
-        await chrome.storage.local.set({ [STRUCTURE_VERSIONS_STORAGE_KEY]: remaining });
-    }, [loadStructureVersions]);
+        await deleteStoredStructureVersion(versionId);
+    }, []);
 
     const loadBookmarkBackups = useCallback(async () => {
-        if (!accountUserId || !canSaveAccountBackups) return [] as BookmarkBackupSnapshot[];
-        try {
-            const res = await fetch(`${BACKEND_URL}/backups/${accountUserId}`, {
-                headers: getAuthHeaders()
-            });
-            if (!res.ok) throw new Error('Failed to load structure backups');
-            const data = await res.json();
-            return data.backups as BookmarkBackupSnapshot[];
-        } catch (err) {
-            console.error('[BACKUPS] Fetch error:', err);
-            return [] as BookmarkBackupSnapshot[];
-        }
-    }, [accountUserId, canSaveAccountBackups, getAuthHeaders]);
+        return backupClient.loadBookmarkBackups();
+    }, [backupClient]);
 
     const saveCurrentBookmarkBackup = useCallback(async (customName?: string) => {
-        if (!accountUserId || !canSaveAccountBackups) throw new Error('Create a free account to save backups.');
-        
-        const name = customName || `Snapshot ${new Date().toLocaleDateString()}`;
-        const res = await fetch(`${BACKEND_URL}/backups/${accountUserId}`, {
-            method: 'POST',
-            headers: buildAuthHeaders(),
-            body: JSON.stringify({ name })
-        });
-
-        if (!res.ok) {
-            const errorData = await res.json().catch(() => ({}));
-            throw new Error(errorData.error || 'Failed to save structure snapshot');
-        }
-        
-        // Return a dummy object just to please TS if it's pushed to local state temporarily
-        return {
-            id: 'just_created',
-            name: name,
-            createdAt: new Date().toISOString(),
-            summary: { folders: 0, bookmarks: 0 }
-        } as BookmarkBackupSnapshot;
-    }, [accountUserId, buildAuthHeaders, canSaveAccountBackups]);
+        return backupClient.saveCurrentBookmarkBackup(customName);
+    }, [backupClient]);
 
     const deleteBookmarkBackup = useCallback(async (backupId: string) => {
-        if (!accountUserId || !canSaveAccountBackups) throw new Error('Create a free account to manage backups.');
-        
-        const res = await fetch(`${BACKEND_URL}/backups/${accountUserId}/${backupId}`, {
-            method: 'DELETE',
-            headers: getAuthHeaders()
-        });
-        
-        if (!res.ok) throw new Error('Failed to delete structure snapshot');
-    }, [accountUserId, canSaveAccountBackups, getAuthHeaders]);
+        await backupClient.deleteBookmarkBackup(backupId);
+    }, [backupClient]);
 
     const restoreBookmarkBackup = useCallback(async (backupId: string) => {
-        if (!accountUserId || !canSaveAccountBackups) throw new Error('Create a free account to restore backups.');
-        
-        const res = await fetch(`${BACKEND_URL}/backups/${accountUserId}/${backupId}/restore`, {
-            method: 'POST',
-            headers: getAuthHeaders()
-        });
-        
-        if (!res.ok) throw new Error('Failed to restore structure snapshot');
-        
-        // Automatically fetch the restored structured results 
-        await fetchResults(accountUserId);
-    }, [accountUserId, canSaveAccountBackups, fetchResults, getAuthHeaders]);
+        await backupClient.restoreBookmarkBackup(backupId);
+        if (accountUserId) {
+            await fetchResults(accountUserId);
+        }
+    }, [accountUserId, backupClient, fetchResults]);
 
     const applyChanges = async () => {
         // Mock mode - just mark as done
@@ -1266,113 +853,9 @@ export const useBookmarkWeaver = (
                 return;
             }
 
-            let movedCount = 0;
-            let skippedCount = 0;
-            let folderCreateFailures = 0;
-            const createdFolderIds = new Map<string, string>();
-            const keepIdsByRoot = new Map<BookmarkRootTitle, Set<string>>();
+            const applyResult = await applyChromeBookmarkPlan(rootNodes);
 
-            const createFoldersForNodes = async (
-                nodes: BookmarkNode[],
-                parentId: string,
-                rootKeepIds: Set<string>,
-                isTopLevel = false
-            ) => {
-                for (const node of nodes) {
-                    if (node.url) continue;
-
-                    try {
-                        const folder = await chrome.bookmarks.create({
-                            parentId,
-                            title: node.title || 'Untitled Folder',
-                        });
-                        createdFolderIds.set(node.id, folder.id);
-                        if (isTopLevel) {
-                            rootKeepIds.add(folder.id);
-                        }
-                        await createFoldersForNodes(node.children || [], folder.id, rootKeepIds);
-                    } catch (error) {
-                        folderCreateFailures += 1;
-                        console.error(`[ApplyChanges] Failed to create folder ${node.title}`, error);
-                    }
-                }
-            };
-
-            const applyBookmarksForNodes = async (
-                nodes: BookmarkNode[],
-                parentId: string,
-                rootKeepIds: Set<string>,
-                isTopLevel = false
-            ) => {
-                for (const node of nodes) {
-                    if (node.url) {
-                        const chromeId = getBookmarkChromeId(node);
-                        if (!chromeId) {
-                            skippedCount += 1;
-                            continue;
-                        }
-
-                        try {
-                            const nextTitle = node.title.trim();
-                            const currentTitle = (node.originalTitle || node.title).trim();
-                            if (nextTitle && nextTitle !== currentTitle) {
-                                await chrome.bookmarks.update(chromeId, { title: node.title });
-                            }
-                            await chrome.bookmarks.move(chromeId, { parentId });
-                            if (isTopLevel) {
-                                rootKeepIds.add(chromeId);
-                            }
-                            movedCount += 1;
-                        } catch (error) {
-                            console.warn(`[ApplyChanges] Failed to move bookmark ${chromeId}:`, error);
-                            skippedCount += 1;
-                        }
-                        continue;
-                    }
-
-                    const folderId = createdFolderIds.get(node.id);
-                    if (!folderId) {
-                        skippedCount += countBookmarksInTree(node.children || []);
-                        continue;
-                    }
-
-                    await applyBookmarksForNodes(node.children || [], folderId, rootKeepIds);
-                }
-            };
-
-            const clearRootChildrenExcept = async (rootId: string, keepIds: Set<string>) => {
-                const children = await chrome.bookmarks.getChildren(rootId);
-                for (const child of [...children].reverse()) {
-                    if (keepIds.has(child.id)) continue;
-                    if (child.url) {
-                        await chrome.bookmarks.remove(child.id);
-                    } else {
-                        await chrome.bookmarks.removeTree(child.id);
-                    }
-                }
-            };
-
-            for (const rootNode of rootNodes) {
-                const rootId = ROOT_IDS[rootNode.rootTitle];
-                const keepIds = new Set<string>();
-                keepIdsByRoot.set(rootNode.rootTitle, keepIds);
-                await createFoldersForNodes(rootNode.children || [], rootId, keepIds, true);
-            }
-
-            for (const rootNode of rootNodes) {
-                const rootId = ROOT_IDS[rootNode.rootTitle];
-                const keepIds = keepIdsByRoot.get(rootNode.rootTitle) || new Set<string>();
-                await applyBookmarksForNodes(rootNode.children || [], rootId, keepIds, true);
-            }
-
-            if (folderCreateFailures === 0 && skippedCount === 0) {
-                for (const rootNode of rootNodes) {
-                    await clearRootChildrenExcept(
-                        ROOT_IDS[rootNode.rootTitle],
-                        keepIdsByRoot.get(rootNode.rootTitle) || new Set<string>()
-                    );
-                }
-            } else {
+            if (applyResult.shouldWarnAboutPartialApply) {
                 window.alert(
                     'Link Loom applied the structure, but some bookmarks could not be moved. Existing folders were left in place to avoid deleting anything unexpectedly.'
                 );
@@ -1384,7 +867,7 @@ export const useBookmarkWeaver = (
             overflowBookmarksRef.current = [];
 
             console.log(
-                `[ApplyChanges] Complete! Moved: ${movedCount}, Skipped: ${skippedCount}, Folder failures: ${folderCreateFailures}`
+                `[ApplyChanges] Complete! Moved: ${applyResult.movedCount}, Skipped: ${applyResult.skippedCount}, Folder failures: ${applyResult.folderCreateFailures}`
             );
             clusterRecoveryTriggered.current = false;
             setErrorMessage(null);
@@ -1405,11 +888,7 @@ export const useBookmarkWeaver = (
         if (!userId) return;
         try {
             await clearPersistedOverflowBookmarks(userId);
-            await fetch(`${BACKEND_URL}/cancel/${userId}`, {
-                method: 'POST',
-                headers: buildAuthHeaders(),
-                body: JSON.stringify({ clearAllQueues: true })
-            });
+            await structureClient.cancel(userId);
         } catch (error) {
             console.error("Cancel error", error);
         } finally {
